+++ src/backend_impls/default_impls.rs
@@ -1,24 +1,37 @@
 use crate::{
     Alignment, F64VectorBackend, I8VectorBackend, I16VectorBackend, I32VectorBackend,
-    I64VectorBackend, I128VectorBackend, IsizeVectorBackend, Length, ScalarDefault,
+    I64VectorBackend, I128VectorBackend, IsizeVectorBackend, Length, ScalarBackend,
     SupportedLength, U8VectorBackend, U16VectorBackend, U32VectorBackend, U64VectorBackend,
     U128VectorBackend, UsizeVectorBackend,
 };
 
-impl ScalarDefault for f64 {}
-impl ScalarDefault for i8 {}
-impl ScalarDefault for i16 {}
-impl ScalarDefault for i32 {}
-impl ScalarDefault for i64 {}
-impl ScalarDefault for i128 {}
-impl ScalarDefault for isize {}
-impl ScalarDefault for u8 {}
-impl ScalarDefault for u16 {}
-impl ScalarDefault for u32 {}
-impl ScalarDefault for u64 {}
-impl ScalarDefault for u128 {}
-impl ScalarDefault for usize {}
-impl ScalarDefault for bool {}
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for f64 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for i8 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for i16 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for i32 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for i64 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for i128 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for isize where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for u8 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for u16 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for u32 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for u64 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for u128 where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for usize where Length<N>: SupportedLength {}
+
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for bool where Length<N>: SupportedLength {}
 
 impl<const N: usize, A: Alignment> F64VectorBackend<N, A> for f64 where Length<N>: SupportedLength {}