+++ src/backend_impls/f32/fallback.rs
@@ -1,5 +1,5 @@
-use crate::{Alignment, F32VectorBackend, Length, ScalarDefault, SupportedLength};
+use crate::{Alignment, F32VectorBackend, Length, ScalarBackend, SupportedLength};
 
-impl ScalarDefault for f32 {}
+impl<const N: usize, A: Alignment> ScalarBackend<N, A> for f32 where Length<N>: SupportedLength {}
 
 impl<const N: usize, A: Alignment> F32VectorBackend<N, A> for f32 where Length<N>: SupportedLength {}