+++ src/backend_impls/f32/sse.rs
@@ -2,55 +2,49 @@
 use core::arch::x86::*;
 #[cfg(target_arch = "x86_64")]
 use core::arch::x86_64::*;
+use core::mem::transmute;
 
-use crate::{
-    Aligned, F32VectorBackend, ScalarBackend, Unaligned, Vec2U, Vec3, Vec3U, Vec4, Vec4U,
-    utils::safe_arch,
-};
+use crate::{Aligned, F32VectorBackend, ScalarBackend, Unaligned, Vec3, Vec4, utils::safe_arch};
 
-unsafe impl ScalarBackend<2, Aligned> for f32 {
-    type VectorRepr = Vec2U<f32>;
-}
+impl ScalarBackend<2, Aligned> for f32 {}
 
-unsafe impl ScalarBackend<3, Aligned> for f32 {
-    type VectorRepr = __m128;
-
+impl ScalarBackend<3, Aligned> for f32 {
     safe_arch! {
         #![target_feature(enable = "sse")]
 
         #[inline]
         fn vec_eq(vec: &Vec3<f32>, other: &Vec3<f32>) -> bool {
-            _mm_movemask_ps(_mm_cmpeq_ps(vec.repr(), other.repr())) as u32 & 0b111 == 0b111
+            _mm_movemask_ps(_mm_cmpeq_ps(vec.to_m128(), other.to_m128())) as u32 & 0b111 == 0b111
         }
 
         #[inline]
         fn vec_ne(vec: &Vec3<f32>, other: &Vec3<f32>) -> bool {
-            _mm_movemask_ps(_mm_cmpneq_ps(vec.repr(), other.repr())) as u32 & 0b111 != 0
+            _mm_movemask_ps(_mm_cmpneq_ps(vec.to_m128(), other.to_m128())) as u32 & 0b111 != 0
         }
 
         #[inline]
         fn vec_neg(vec: Vec3<f32>) -> Vec3<f32> {
-            neg(vec.repr()).to_vec3()
+            neg(vec.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_add(vec: Vec3<f32>, rhs: Vec3<f32>) -> Vec3<f32> {
-            _mm_add_ps(vec.repr(), rhs.repr()).to_vec3()
+            _mm_add_ps(vec.to_m128(), rhs.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_sub(vec: Vec3<f32>, rhs: Vec3<f32>) -> Vec3<f32> {
-            _mm_sub_ps(vec.repr(), rhs.repr()).to_vec3()
+            _mm_sub_ps(vec.to_m128(), rhs.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_mul(vec: Vec3<f32>, rhs: Vec3<f32>) -> Vec3<f32> {
-            _mm_mul_ps(vec.repr(), rhs.repr()).to_vec3()
+            _mm_mul_ps(vec.to_m128(), rhs.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_div(vec: Vec3<f32>, rhs: Vec3<f32>) -> Vec3<f32> {
-            _mm_div_ps(vec.repr(), rhs.repr()).to_vec3()
+            _mm_div_ps(vec.to_m128(), rhs.to_m128()).to_vec3()
         }
     }
 
@@ -60,50 +54,48 @@ unsafe impl ScalarBackend<3, Aligned> for f32 {
 
         #[inline]
         fn vec_rem(vec: Vec3<f32>, rhs: Vec3<f32>) -> Vec3<f32> {
-            rem(vec.repr(), rhs.repr()).to_vec3()
+            rem(vec.to_m128(), rhs.to_m128()).to_vec3()
         }
     }
 }
 
-unsafe impl ScalarBackend<4, Aligned> for f32 {
-    type VectorRepr = __m128;
-
+impl ScalarBackend<4, Aligned> for f32 {
     safe_arch! {
         #![target_feature(enable = "sse")]
 
         #[inline]
         fn vec_eq(vec: &Vec4<f32>, other: &Vec4<f32>) -> bool {
-            _mm_movemask_ps(_mm_cmpeq_ps(vec.repr(), other.repr())) as u32 == 0xf
+            _mm_movemask_ps(_mm_cmpeq_ps(vec.to_m128(), other.to_m128())) as u32 == 0xf
         }
 
         #[inline]
         fn vec_ne(vec: &Vec4<f32>, other: &Vec4<f32>) -> bool {
-            _mm_movemask_ps(_mm_cmpneq_ps(vec.repr(), other.repr())) as u32 != 0
+            _mm_movemask_ps(_mm_cmpneq_ps(vec.to_m128(), other.to_m128())) as u32 != 0
         }
 
         #[inline]
         fn vec_neg(vec: Vec4<f32>) -> Vec4<f32> {
-            neg(vec.repr()).to_vec4()
+            neg(vec.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_add(vec: Vec4<f32>, rhs: Vec4<f32>) -> Vec4<f32> {
-            _mm_add_ps(vec.repr(), rhs.repr()).to_vec4()
+            _mm_add_ps(vec.to_m128(), rhs.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_sub(vec: Vec4<f32>, rhs: Vec4<f32>) -> Vec4<f32> {
-            _mm_sub_ps(vec.repr(), rhs.repr()).to_vec4()
+            _mm_sub_ps(vec.to_m128(), rhs.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_mul(vec: Vec4<f32>, rhs: Vec4<f32>) -> Vec4<f32> {
-            _mm_mul_ps(vec.repr(), rhs.repr()).to_vec4()
+            _mm_mul_ps(vec.to_m128(), rhs.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_div(vec: Vec4<f32>, rhs: Vec4<f32>) -> Vec4<f32> {
-            _mm_div_ps(vec.repr(), rhs.repr()).to_vec4()
+            _mm_div_ps(vec.to_m128(), rhs.to_m128()).to_vec4()
         }
     }
 
@@ -113,22 +105,14 @@ unsafe impl ScalarBackend<4, Aligned> for f32 {
 
         #[inline]
         fn vec_rem(vec: Vec4<f32>, rhs: Vec4<f32>) -> Vec4<f32> {
-            rem(vec.repr(), rhs.repr()).to_vec4()
+            rem(vec.to_m128(), rhs.to_m128()).to_vec4()
         }
     }
 }
 
-unsafe impl ScalarBackend<2, Unaligned> for f32 {
-    type VectorRepr = Vec2U<f32>;
-}
-
-unsafe impl ScalarBackend<3, Unaligned> for f32 {
-    type VectorRepr = Vec3U<f32>;
-}
-
-unsafe impl ScalarBackend<4, Unaligned> for f32 {
-    type VectorRepr = Vec4U<f32>;
-}
+impl ScalarBackend<2, Unaligned> for f32 {}
+impl ScalarBackend<3, Unaligned> for f32 {}
+impl ScalarBackend<4, Unaligned> for f32 {}
 
 impl F32VectorBackend<2, Aligned> for f32 {}
 
@@ -138,42 +122,42 @@ impl F32VectorBackend<3, Aligned> for f32 {
 
         #[inline]
         fn vec_is_nan(vec: Vec3<f32>) -> bool {
-            _mm_movemask_ps(nan_mask(vec.repr())) as u32 & 0b111 != 0
+            _mm_movemask_ps(nan_mask(vec.to_m128())) as u32 & 0b111 != 0
         }
 
         #[inline]
         fn vec_is_finite(vec: Vec3<f32>) -> bool {
-            _mm_movemask_ps(finite_mask(vec.repr())) as u32 & 0b111 == 0b111
+            _mm_movemask_ps(finite_mask(vec.to_m128())) as u32 & 0b111 == 0b111
         }
 
         #[inline]
         fn vec_max(vec: Vec3<f32>, other: Vec3<f32>) -> Vec3<f32> {
-            _mm_max_ps(vec.repr(), other.repr()).to_vec3()
+            _mm_max_ps(vec.to_m128(), other.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_min(vec: Vec3<f32>, other: Vec3<f32>) -> Vec3<f32> {
-            _mm_min_ps(vec.repr(), other.repr()).to_vec3()
+            _mm_min_ps(vec.to_m128(), other.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_abs(vec: Vec3<f32>) -> Vec3<f32> {
-            abs(vec.repr()).to_vec3()
+            abs(vec.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_signum(vec: Vec3<f32>) -> Vec3<f32> {
-            signum(vec.repr()).to_vec3()
+            signum(vec.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_copysign(vec: Vec3<f32>, sign: Vec3<f32>) -> Vec3<f32> {
-            copysign(vec.repr(), sign.repr()).to_vec3()
+            copysign(vec.to_m128(), sign.to_m128()).to_vec3()
         }
 
         #[inline]
         fn vec_element_sum(vec: Vec3<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_add_ps(vec, _mm_shuffle_ps(vec, _mm_set1_ps(0.0), 0b00_11_00_01));
             let vec = _mm_add_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_10));
             _mm_cvtss_f32(vec)
@@ -181,7 +165,7 @@ impl F32VectorBackend<3, Aligned> for f32 {
 
         #[inline]
         fn vec_element_product(vec: Vec3<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_mul_ps(vec, _mm_shuffle_ps(vec, _mm_set1_ps(1.0), 0b00_11_00_01));
             let vec = _mm_mul_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_10));
             _mm_cvtss_f32(vec)
@@ -189,7 +173,7 @@ impl F32VectorBackend<3, Aligned> for f32 {
 
         #[inline]
         fn vec_max_element(vec: Vec3<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_max_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_10_10));
             let vec = _mm_max_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_01));
             _mm_cvtss_f32(vec)
@@ -197,7 +181,7 @@ impl F32VectorBackend<3, Aligned> for f32 {
 
         #[inline]
         fn vec_min_element(vec: Vec3<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_min_ps(vec, _mm_shuffle_ps(vec, vec, 0b01_01_10_10));
             let vec = _mm_min_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_01));
             _mm_cvtss_f32(vec)
@@ -206,13 +190,13 @@ impl F32VectorBackend<3, Aligned> for f32 {
         #[cfg(backend)]
         #[inline(always)]
         fn vec_round(vec: Vec3<f32>) -> Vec3<f32> {
-            round(vec.repr()).to_vec3()
+            round(vec.to_m128()).to_vec3()
         }
 
         #[cfg(backend)]
         #[inline(always)]
         fn vec_sqrt(vec: Vec3<f32>) -> Vec3<f32> {
-            _mm_sqrt_ps(vec.repr()).to_vec3()
+            _mm_sqrt_ps(vec.to_m128()).to_vec3()
         }
     }
 
@@ -222,19 +206,19 @@ impl F32VectorBackend<3, Aligned> for f32 {
         #[cfg(backend)]
         #[inline(always)]
         fn vec_floor(vec: Vec3<f32>) -> Vec3<f32> {
-            floor(vec.repr()).to_vec3()
+            floor(vec.to_m128()).to_vec3()
         }
 
         #[cfg(backend)]
         #[inline(always)]
         fn vec_ceil(vec: Vec3<f32>) -> Vec3<f32> {
-            ceil(vec.repr()).to_vec3()
+            ceil(vec.to_m128()).to_vec3()
         }
 
         #[cfg(backend)]
         #[inline(always)]
         fn vec_trunc(vec: Vec3<f32>) -> Vec3<f32> {
-            trunc(vec.repr()).to_vec3()
+            trunc(vec.to_m128()).to_vec3()
         }
     }
 }
@@ -245,42 +229,42 @@ impl F32VectorBackend<4, Aligned> for f32 {
 
         #[inline]
         fn vec_is_nan(vec: Vec4<f32>) -> bool {
-            _mm_movemask_ps(nan_mask(vec.repr())) as u32 != 0
+            _mm_movemask_ps(nan_mask(vec.to_m128())) as u32 != 0
         }
 
         #[inline]
         fn vec_is_finite(vec: Vec4<f32>) -> bool {
-            _mm_movemask_ps(finite_mask(vec.repr())) as u32 == 0xf
+            _mm_movemask_ps(finite_mask(vec.to_m128())) as u32 == 0xf
         }
 
         #[inline]
         fn vec_max(vec: Vec4<f32>, other: Vec4<f32>) -> Vec4<f32> {
-            _mm_max_ps(vec.repr(), other.repr()).to_vec4()
+            _mm_max_ps(vec.to_m128(), other.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_min(vec: Vec4<f32>, other: Vec4<f32>) -> Vec4<f32> {
-            _mm_min_ps(vec.repr(), other.repr()).to_vec4()
+            _mm_min_ps(vec.to_m128(), other.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_abs(vec: Vec4<f32>) -> Vec4<f32> {
-            abs(vec.repr()).to_vec4()
+            abs(vec.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_signum(vec: Vec4<f32>) -> Vec4<f32> {
-            signum(vec.repr()).to_vec4()
+            signum(vec.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_copysign(vec: Vec4<f32>, sign: Vec4<f32>) -> Vec4<f32> {
-            copysign(vec.repr(), sign.repr()).to_vec4()
+            copysign(vec.to_m128(), sign.to_m128()).to_vec4()
         }
 
         #[inline]
         fn vec_element_sum(vec: Vec4<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_add_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_11_00_01));
             let vec = _mm_add_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_10));
             _mm_cvtss_f32(vec)
@@ -288,7 +272,7 @@ impl F32VectorBackend<4, Aligned> for f32 {
 
         #[inline]
         fn vec_element_product(vec: Vec4<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_mul_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_11_00_01));
             let vec = _mm_mul_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_10));
             _mm_cvtss_f32(vec)
@@ -296,7 +280,7 @@ impl F32VectorBackend<4, Aligned> for f32 {
 
         #[inline]
         fn vec_max_element(vec: Vec4<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_max_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_11_10));
             let vec = _mm_max_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_01));
             _mm_cvtss_f32(vec)
@@ -304,7 +288,7 @@ impl F32VectorBackend<4, Aligned> for f32 {
 
         #[inline]
         fn vec_min_element(vec: Vec4<f32>) -> f32 {
-            let vec = vec.repr();
+            let vec = vec.to_m128();
             let vec = _mm_min_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_11_10));
             let vec = _mm_min_ps(vec, _mm_shuffle_ps(vec, vec, 0b00_00_00_01));
             _mm_cvtss_f32(vec)
@@ -313,13 +297,13 @@ impl F32VectorBackend<4, Aligned> for f32 {
         #[cfg(backend)]
         #[inline(always)]
         fn vec_round(vec: Vec4<f32>) -> Vec4<f32> {
-            round(vec.repr()).to_vec4()
+            round(vec.to_m128()).to_vec4()
         }
 
         #[cfg(backend)]
         #[inline(always)]
         fn vec_sqrt(vec: Vec4<f32>) -> Vec4<f32> {
-            _mm_sqrt_ps(vec.repr()).to_vec4()
+            _mm_sqrt_ps(vec.to_m128()).to_vec4()
         }
     }
 
@@ -329,19 +313,19 @@ impl F32VectorBackend<4, Aligned> for f32 {
         #[cfg(backend)]
         #[inline(always)]
         fn vec_floor(vec: Vec4<f32>) -> Vec4<f32> {
-            floor(vec.repr()).to_vec4()
+            floor(vec.to_m128()).to_vec4()
         }
 
         #[cfg(backend)]
         #[inline(always)]
         fn vec_ceil(vec: Vec4<f32>) -> Vec4<f32> {
-            ceil(vec.repr()).to_vec4()
+            ceil(vec.to_m128()).to_vec4()
         }
 
         #[cfg(backend)]
         #[inline(always)]
         fn vec_trunc(vec: Vec4<f32>) -> Vec4<f32> {
-            trunc(vec.repr()).to_vec4()
+            trunc(vec.to_m128()).to_vec4()
         }
     }
 }
@@ -483,21 +467,35 @@ fn trunc(vec: __m128) -> __m128 {
     )
 }
 
+trait ToM128 {
+    fn to_m128(self) -> __m128;
+}
+
 trait ToVec {
     fn to_vec3(self) -> Vec3<f32>;
     fn to_vec4(self) -> Vec4<f32>;
 }
 
+impl ToM128 for Vec3<f32> {
+    fn to_m128(self) -> __m128 {
+        unsafe { transmute::<Vec3<f32>, __m128>(self) }
+    }
+}
+
+impl ToM128 for Vec4<f32> {
+    fn to_m128(self) -> __m128 {
+        unsafe { transmute::<Vec4<f32>, __m128>(self) }
+    }
+}
+
 impl ToVec for __m128 {
     #[inline]
     fn to_vec3(self) -> Vec3<f32> {
-        // SAFETY: any value of `__m128` is a valid value of `Vec3<f32>`.
-        unsafe { Vec3::from_repr(self) }
+        unsafe { transmute::<__m128, Vec3<f32>>(self) }
     }
 
     #[inline]
     fn to_vec4(self) -> Vec4<f32> {
-        // SAFETY: any value of `__m128` is a valid value of `Vec4<f32>`.
-        unsafe { Vec4::from_repr(self) }
+        unsafe { transmute::<__m128, Vec4<f32>>(self) }
     }
 }