+++ tests/mod.rs
@@ -1,8 +1,7 @@
 #![allow(clippy::bool_assert_comparison)]
 
-mod default_scalar;
 mod integrations;
 mod mask;
-mod simd_scalar;
+mod scalar;
 mod utils;
 mod vector;